decision-tree-automation/decision-tree-automation-api/requirements.txt
fastapi
uvicorn
requests
apscheduler
psycopg2-binary
sqlalchemy
python-dotenv
pytz
